model/requirements.txt
# Core dependencies
pypdf2==3.0.1
langchain==0.3.27
python-dotenv==1.1.1
langchain-community==0.3.30
langchain-openai==0.3.34

# API Framework
flask==3.0.0
flask-cors==4.0.0

# Vector Database (ChromaDB instead of FAISS)
chromadb==0.4.22

# LLM Providers
openai==1.109.1
tiktoken==0.11.0
langchain-anthropic==0.3.22
langchain-google-genai==2.0.10
google-generativeai==0.8.5

# Optional: Local embeddings (uncomment if needed)
# sentence-transformers==5.1.1
# torch==2.8.0

